+++ src/Microsoft.AspNet.Mvc.Core/Routing/AttributeRouteLinkGenerationEntry.cs
@@ -8,8 +8,8 @@ using Microsoft.AspNet.Routing.Template;
 namespace Microsoft.AspNet.Mvc.Routing
 {
     /// <summary>
-    /// Used to build an <see cref="AttributeRoute"/>. Represents an individual URL-generating route that will be
-    /// aggregated into the <see cref="AttributeRoute"/>.
+    /// Used to build an <see cref="InnerAttributeRoute"/>. Represents an individual URL-generating route that will be
+    /// aggregated into the <see cref="InnerAttributeRoute"/>.
     /// </summary>
     public class AttributeRouteLinkGenerationEntry
     {
